Chuẩn hoá response trả về

+++ gearshop-backend/src/main/java/com/uchiha/gearshop/dao/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     @Query(value = "select u from UserEntity u where u.username = :username")
     UserEntity findByUsername(String username);
 }
