+++ gearshop-backend/src/main/java/com/uchiha/gearshop/service/UserService.java
@@ -12,7 +12,7 @@ public interface UserService {
 
     int update(UserRequestDTO userRequestDTO);
 
-    int delete(Long id);
+    int delete(int id);
 
-    UserEntity getById(Long id);
+    UserEntity getById(int id);
 }
